Refactor shared path and span helpers
- extract span, path, and expression helpers into dedicated modules\n- reuse a generic Path<String> for attribute parsing with empty segment filtering\n- document the design update and clarify the main-function detection limitation
