+++ common/src/attributes/tests.rs
@@ -1,3 +1,5 @@
+//! Tests for attribute helpers.
+
 use super::*;
 use rstest::rstest;
