+++ common/src/lib.rs
@@ -1,8 +1,12 @@
 #![deny(clippy::expect_used, clippy::unwrap_used)]
 
+//! Shared utilities for Whitaker lint implementations.
+
 pub mod attributes;
 pub mod context;
 pub mod diagnostics;
+pub mod expr;
+pub mod path;
 pub mod span;
 
 pub use attributes::{
@@ -11,7 +15,6 @@ pub use attributes::{
 };
 pub use context::{ContextEntry, ContextKind, in_test_like_context, is_in_main_fn, is_test_fn};
 pub use diagnostics::{Applicability, Diagnostic, DiagnosticBuilder, Suggestion, span_lint};
-pub use span::{
-    Expr, SimplePath, SourceLocation, SourceSpan, SpanError, def_id_of_expr_callee, is_path_to,
-    module_line_count, recv_is_option_or_result, span_to_lines,
-};
+pub use expr::{Expr, def_id_of_expr_callee, is_path_to, recv_is_option_or_result};
+pub use path::{Path, SimplePath};
+pub use span::{SourceLocation, SourceSpan, SpanError, module_line_count, span_to_lines};
