+++ common/src/span.rs
@@ -1,3 +1,5 @@
+//! Utilities for working with source locations and spans.
+
 use std::ops::RangeInclusive;
 
 /// Errors produced when constructing spans.
@@ -64,9 +66,10 @@ impl SourceSpan {
     #[must_use]
     pub fn new(start: SourceLocation, end: SourceLocation) -> Result<Self, SpanError> {
         if start.line > end.line || (start.line == end.line && start.column > end.column) {
-            return Err(SpanError::StartAfterEnd);
+            Err(SpanError::StartAfterEnd)
+        } else {
+            Ok(Self { start, end })
         }
-        Ok(Self { start, end })
     }
 
     /// Returns the start location.
@@ -112,117 +115,6 @@ pub fn module_line_count(span: SourceSpan) -> usize {
     span.end.line() - span.start.line() + 1
 }
 
-/// A simplified representation of a Rust path.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct SimplePath {
-    segments: Vec<String>,
-}
-
-impl SimplePath {
-    /// Builds a path from iterator segments.
-    #[must_use]
-    pub fn new<I, S>(segments: I) -> Self
-    where
-        I: IntoIterator<Item = S>,
-        S: Into<String>,
-    {
-        Self {
-            segments: segments.into_iter().map(Into::into).collect(),
-        }
-    }
-
-    /// Parses a path from the conventional `::`-separated string form.
-    #[must_use]
-    pub fn from(path: &str) -> Self {
-        Self::new(path.split("::"))
-    }
-
-    /// Returns the path segments.
-    #[must_use]
-    pub fn segments(&self) -> &[String] {
-        &self.segments
-    }
-
-    /// Returns `true` when the path matches the provided segments exactly.
-    #[must_use]
-    pub fn matches<'a, I>(&self, candidate: I) -> bool
-    where
-        I: IntoIterator<Item = &'a str>,
-    {
-        self.segments
-            .iter()
-            .map(String::as_str)
-            .eq(candidate.into_iter())
-    }
-}
-
-/// A tiny expression model used for helper functions.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum Expr {
-    /// A call expression with a resolved callee path.
-    Call { callee: SimplePath },
-    /// A path expression.
-    Path(SimplePath),
-    /// Any other literal expression (placeholder for expansion).
-    Literal(String),
-}
-
-/// Returns the callee path of a call expression, if one is present.
-///
-/// # Examples
-///
-/// ```
-/// use common::span::{def_id_of_expr_callee, Expr, SimplePath};
-///
-/// let expr = Expr::Call { callee: SimplePath::from("std::mem::drop") };
-/// assert_eq!(def_id_of_expr_callee(&expr).unwrap().segments(), &["std", "mem", "drop"]);
-/// ```
-#[must_use]
-pub fn def_id_of_expr_callee(expr: &Expr) -> Option<&SimplePath> {
-    match expr {
-        Expr::Call { callee } => Some(callee),
-        _ => None,
-    }
-}
-
-/// Tests whether a path matches the provided candidate segments.
-///
-/// # Examples
-///
-/// ```
-/// use common::span::{is_path_to, SimplePath};
-///
-/// let path = SimplePath::from("core::option::Option");
-/// assert!(is_path_to(&path, ["core", "option", "Option"]));
-/// ```
-#[must_use]
-pub fn is_path_to<'a, I>(path: &SimplePath, candidate: I) -> bool
-where
-    I: IntoIterator<Item = &'a str>,
-{
-    path.matches(candidate)
-}
-
-/// Returns `true` when the receiver is `Option` or `Result` regardless of
-/// module path.
-///
-/// # Examples
-///
-/// ```
-/// use common::span::{recv_is_option_or_result, SimplePath};
-///
-/// assert!(recv_is_option_or_result(&SimplePath::from("std::option::Option")));
-/// assert!(recv_is_option_or_result(&SimplePath::from("Result")));
-/// assert!(!recv_is_option_or_result(&SimplePath::from("crate::Thing")));
-/// ```
-#[must_use]
-pub fn recv_is_option_or_result(path: &SimplePath) -> bool {
-    matches!(
-        path.segments().last().map(String::as_str),
-        Some("Option" | "Result")
-    )
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -239,31 +131,5 @@ mod tests {
         let span = SourceSpan::new(SourceLocation::new(5, 0), SourceLocation::new(7, 3)).unwrap();
         assert_eq!(span_to_lines(span), 5..=7);
         assert_eq!(module_line_count(span), 3);
-    }
-
-    #[rstest]
-    fn callee_extraction() {
-        let expr = Expr::Call {
-            callee: SimplePath::from("std::mem::drop"),
-        };
-        assert!(def_id_of_expr_callee(&expr).is_some());
-    }
-
-    #[rstest]
-    fn recognises_option_like_receivers() {
-        let option_path = SimplePath::from("std::option::Option");
-        let result_path = SimplePath::from("Result");
-        let custom_path = SimplePath::from("crate::Thing");
-
-        assert!(recv_is_option_or_result(&option_path));
-        assert!(recv_is_option_or_result(&result_path));
-        assert!(!recv_is_option_or_result(&custom_path));
-    }
-
-    #[rstest]
-    fn path_comparison() {
-        let path = SimplePath::from("crate::module::Item");
-        assert!(is_path_to(&path, ["crate", "module", "Item"]));
-        assert!(!is_path_to(&path, ["crate", "module", "Other"]));
     }
 }
